# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cloud-Ready School Management ERP API - Phase 1",
        "description": "Complete RESTful API specifications for Multi-Campus School Management ERP. Encompasses Institute Management, Multi-Campus Branches, Academic Years & Terms, Spatie Role-Based Access Control (RBAC), Global Settings, Classes, Sections, Streams, Houses, Subjects, Employee & Staff Management, and Student Admissions & Profiles.",
        "version": "1.0.0",
        "contact": {
            "name": "School ERP Support",
            "email": "support@schoolerp.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Laravel Development API Server"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Sanctum API Bearer Token"
            }
        }
    },
    "paths": {
...
        "/academic-years": {
            "get": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List Academic Years",
                "responses": {
                    "200": {
                        "description": "List of academic years"
                    }
                }
            },
            "post": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create Academic Year",
                "responses": {
                    "201": {
                        "description": "Academic year created"
                    }
                }
            }
        },
        "/academic-years/{id}": {
            "get": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get Academic Year Details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Academic year details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Update Academic Year",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Academic year updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Delete Academic Year",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Academic year deleted"
                    }
                }
            }
        },
        "/academic-years/{id}/set-current": {
            "post": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Set as Active Session",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Academic year set as current active session"
                    }
                }
            }
        },
        "/academic-years/{id}/terms": {
            "post": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Add Term / Semester",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Term created"
                    }
                }
            }
        },
        "/terms/{termId}": {
            "put": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Update Term",
                "parameters": [
                    {
                        "name": "termId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Term updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Delete Term",
                "parameters": [
                    {
                        "name": "termId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Term deleted"
                    }
                }
            }
        },
        "/academic-years/{id}/promotion-rules": {
            "put": {
                "tags": [
                    "Academic Years & Sessions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Configure Promotion Rules",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promotion rules configured"
                    }
                }
            }
        },
...
    }
}